Reduce duplication and get dimensions by stateCode
