+++ server/collections/BaseMetrics.js
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-const { default: RESOURCES } = require("./resources");
-const { COLLECTIONS } = require("./resources/shared");
+const { COLLECTIONS } = require("../constants/collections");
+const { default: collectionsByStateCode } = require("./resources");
 /**
  * The base class for all metrics. Use the helper `getResourcesByType` to instantiate a metric
  * by metricType and stateCode.
@@ -31,7 +31,7 @@ class BaseMetrics {
     this.constructor.validateMetricType(metricType);
     this.stateCode = stateCode;
     this.metricType = metricType;
-    this.metrics = RESOURCES[stateCode][metricType];
+    this.metrics = collectionsByStateCode[stateCode][metricType];
   }
 
   static validateMetricType(metricType) {
