+++ server/collections/resources/index.js
@@ -1,12 +1,9 @@
-const US_MO = require("./US_MO");
-const US_PA = require("./US_PA");
-const US_ND = require("./US_ND");
-
-const US_DEMO = { ...US_MO, ...US_ND };
+const { default: getCollections } = require("./getCollections");
+const { US_MO, US_PA, US_ND } = require("../../constants/stateCodes");
 
 exports.default = {
-  US_ND,
-  US_MO,
-  US_PA,
-  US_DEMO,
+  US_MO: getCollections(US_MO),
+  US_PA: getCollections(US_PA),
+  US_DEMO: getCollections(US_MO),
+  US_ND: getCollections(US_ND),
 };
