+++ server/constants/collections.js
@@ -1,0 +1,12 @@
+const COLLECTIONS = {
+  NEW_REVOCATION: "newRevocation",
+  COMMUNITY_GOALS: "communityGoals",
+  COMMUNITY_EXPLORE: "communityExplore",
+  FACILITIES_GOALS: "facilitiesGoals",
+  FACILITIES_EXPLORE: "facilitiesExplore",
+  PROGRAMMING_EXPLORE: "programmingExplore",
+};
+
+module.exports = {
+  COLLECTIONS,
+};
