+++ server/constants/stateCodes.js
@@ -1,0 +1,5 @@
+module.exports = {
+  US_MO: "US_MO",
+  US_PA: "US_PA",
+  US_ND: "US_ND",
+};
